% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Obtain compilation flags for \code{Rmstoolkitlib}}
\source{
This function (mechanism) was based on zlibbioc (https://github.com/Bioconductor/zlibbioc)
}
\usage{
pkgconfig(opt = c("PKG_CFLAGS", "PKG_CXXFLAGS", "PKG_LIBS"))
}
\arguments{
\item{opt}{A character string specifying which compilation flags to return.
Must be one of \code{"PKG_CFLAGS"}, \code{"PKG_CXXFLAGS"}, or \code{"PKG_LIBS"}.}
}
\value{
A character string containing the requested compilation flags.
}
\description{
Obtain compilation flags for \code{Rmstoolkitlib}
}
